feat(protocol-designer): allow file upload
- DRY out configureStore
