+++ protocol-designer/src/file-data/reducers/index.js
@@ -2,8 +2,14 @@
 import {combineReducers} from 'redux'
 import {handleActions, type ActionType} from 'redux-actions'
 
-import {updateFileMetadataFields, saveFileMetadata} from '../actions'
+import {actions as loadFileActions} from '../../load-file'
+import {
+  saveFileMetadata,
+  updateFileMetadataFields
+} from '../actions'
 import type {FileMetadataFields} from '../types'
+
+const {LOAD_FILE, loadFile} = loadFileActions
 
 const defaultFields = {
   name: '',
@@ -12,6 +18,7 @@ const defaultFields = {
 }
 
 const unsavedMetadataForm = handleActions({
+  [LOAD_FILE]: () => defaultFields,
   UPDATE_FILE_METADATA_FIELDS: (state: FileMetadataFields, action: ActionType<typeof updateFileMetadataFields>) => ({
     ...state,
     ...action.payload
@@ -23,6 +30,14 @@ const unsavedMetadataForm = handleActions({
 }, defaultFields)
 
 const fileMetadata = handleActions({
+  [LOAD_FILE]: (state: FileMetadataFields, action: ActionType<typeof loadFile>) => {
+    const {metadata} = action.payload
+    return {
+      author: metadata.author,
+      description: metadata.description,
+      name: metadata['protocol-name']
+    }
+  },
   SAVE_FILE_METADATA: (state: FileMetadataFields, action: ActionType<typeof saveFileMetadata>) => ({
     ...state,
     ...action.payload
