+++ protocol-designer/src/load-file/actions.js
@@ -1,0 +1,9 @@
+// @flow
+import type {ProtocolFile} from '../file-types'
+
+export const LOAD_FILE: 'LOAD_FILE' = 'LOAD_FILE'
+// expects valid, parsed JSON protocol.
+export const loadFile = (payload: ProtocolFile) => ({
+  type: LOAD_FILE,
+  payload
+})
