+++ protocol-designer/src/load-file/index.js
@@ -1,0 +1,6 @@
+// @flow
+import * as actions from './actions'
+
+export {
+  actions
+}
